Move tensor attribute source to separate directory.
