+++ searchcore/src/tests/proton/docsummary/docsummary.cpp
@@ -23,7 +23,7 @@ LOG_SETUP("docsummary_test");
 #include <vespa/searchlib/transactionlog/nosyncproxy.h>
 #include <vespa/vespalib/tensor/tensor_factory.h>
 #include <vespa/vespalib/tensor/default_tensor.h>
-#include <vespa/searchlib/attribute/tensorattribute.h>
+#include <vespa/searchlib/tensor/tensor_attribute.h>
 
 using namespace document;
 using namespace search;
