+++ searchlib/src/vespa/searchlib/attribute/createsinglestd.cpp
@@ -11,7 +11,7 @@ LOG_SETUP(".createsinglestd");
 #include <vespa/searchlib/attribute/attributevector.hpp>
 #include <vespa/searchlib/attribute/singlenumericattribute.hpp>
 #include <vespa/searchlib/attribute/singlestringattribute.h>
-#include <vespa/searchlib/attribute/tensorattribute.h>
+#include <vespa/searchlib/tensor/tensor_attribute.h>
 
 namespace search {
