+++ searchlib/src/vespa/searchlib/tensor/tensor_attribute.cpp
@@ -1,10 +1,10 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include <vespa/fastos/fastos.h>
-#include "tensorattribute.h"
+#include "tensor_attribute.h"
 #include <vespa/vespalib/tensor/default_tensor.h>
 #include <vespa/vespalib/tensor/tensor.h>
-#include "tensorattributesaver.h"
+#include "tensor_attribute_saver.h"
 
 using vespalib::eval::ValueType;
 using vespalib::tensor::Tensor;
