+++ searchlib/src/vespa/searchlib/tensor/tensor_attribute_saver.cpp
@@ -1,7 +1,7 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include <vespa/fastos/fastos.h>
-#include "tensorattributesaver.h"
+#include "tensor_attribute_saver.h"
 #include <vespa/searchlib/util/bufferwriter.h>
 
 using vespalib::GenerationHandler;
