+++ searchlib/src/vespa/searchlib/tensor/tensor_store.cpp
@@ -1,7 +1,7 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include <vespa/fastos/fastos.h>
-#include "tensorstore.h"
+#include "tensor_store.h"
 #include <vespa/vespalib/tensor/tensor.h>
 #include <vespa/vespalib/tensor/serialization/typed_binary_format.h>
 #include <vespa/vespalib/objects/nbostream.h>
